refactor: rename multiOperationSpinner filename
